+++ src/scripts/SafetyCardName.js
@@ -1,0 +1,10 @@
+export var SafetyCardName;
+(function (SafetyCardName) {
+  SafetyCardName["X"] = "X";
+  SafetyCardName["O"] = "O";
+  SafetyCardName["N"] = "N";
+  SafetyCardName["Pause"] = "Pause";
+  SafetyCardName["Resume"] = "Resume";
+  SafetyCardName["FastForward"] = "FastForward";
+  SafetyCardName["Rewind"] = "Rewind";
+})(SafetyCardName || (SafetyCardName = {}));
