--- src/scripts/SafetyCardName.ts
@@ -1,9 +1,0 @@
-export enum SafetyCardName {
-	X = "X",
-	O = "O",
-	N = "N",
-	Pause = "Pause",
-	Resume = "Resume",
-	FastForward = "FastForward",
-	Rewind = "Rewind"
-}
