+++ src/scripts/constants.js
@@ -1,8 +1,6 @@
 const CONSTANTS = {
-	MODULE_NAME: "safety-tools",
-	PATH: `modules/safety-tools/`
+  MODULE_NAME: "safety-tools",
+  PATH: `modules/safety-tools/`,
 };
-
 CONSTANTS.PATH = `modules/${CONSTANTS.MODULE_NAME}/`;
-
 export default CONSTANTS;
